+++ src/main/java/eccs/com/modules/controlventas/catalogoclientes/service/CatalogoClientesService.java
@@ -1,0 +1,7 @@
+package eccs.com.modules.controlventas.catalogoclientes.service;
+
+import eccs.com.core.dtos.ResponseDto;
+
+public interface CatalogoClientesService {
+    ResponseDto<Object> getCatalogoClientes();
+}